rename: 홈 컴포넌트 폴더 (home)>components 에서 componets로 변경
