+++ src/app/components/Category/CategoryButton.tsx
@@ -10,7 +10,7 @@ const CategoryButton = ({ children, className, onClick }: CategoryButtonProps) =
   return (
     <button
       className={clsx(
-        'flex items-center justify-center rounded-[16px] border border-gray-300 bg-gray-100 transition-colors hover:bg-gray-200 hover:text-black',
+        'flex items-center justify-center rounded-[16px] border border-gray-300 bg-gray-100 transition-colors hover:bg-gray-200',
         className,
       )}
       onClick={onClick}
